requirements/base.txt
django==4.1.4
django-environ==0.9.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-filter==22.1

Pillow==9.0.0

argon2-cffi==21.3.0

djoser==2.1.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.6.0